relationships established between user and contact,,protected all public routes

+++ Controllers/contact_controller.js
@@ -4,7 +4,7 @@ const asyncHandler = require("express-async-handler");
 const contactModel = require("../Models/contactModel");
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await contactModel.find();
+    const contacts = await contactModel.find({ user_id: req.user.id });
     return res.status(200).json(contacts);
 });
 
@@ -17,16 +17,16 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await contactModel.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id,
     });
     return res.status(201).json(contact);
 });
-    
 
 const getContact = asyncHandler(async (req, res) => {
     const id = req.params.id;
     const contact = await contactModel.findById(id);
-    if(!contact){
+    if (!contact) {
         res.status(404);
         throw new Error("Contact Not Found");
     }
@@ -36,9 +36,14 @@ const getContact = asyncHandler(async (req, res) => {
 const updateContact = asyncHandler(async (req, res) => {
     const id = req.params.id;
     const contact = await contactModel.findById(id);
-    if(!contact){
+    if (!contact) {
         res.status(404);
         throw new Error("Contact Not Found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
     }
 
     const updatedContact = await contactModel.findByIdAndUpdate(
@@ -52,9 +57,14 @@ const updateContact = asyncHandler(async (req, res) => {
 const deleteContact = asyncHandler(async (req, res) => {
     const id = req.params.id;
     const contact = await contactModel.findById(id);
-    if(!contact){
+    if (!contact) {
         res.status(404);
         throw new Error("Contact Not Found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to delete other user contacts");
     }
 
     await contactModel.deleteOne({ _id: id });
